+++ craco.config.cjs
@@ -1,14 +1,12 @@
-const locatorPlugin = require("@locator/babel-jsx").default;
-
 module.exports = {
-  babel: {
-    plugins: [
-      [
-        locatorPlugin,
-        {
-          env: "development",
-        },
-      ],
-    ],
-  },
+  // babel: {
+  //   plugins: [
+  //     [
+  //       locatorPlugin,
+  //       {
+  //         env: "development",
+  //       },
+  //     ],
+  //   ],
+  // },
 };